+++ lib/middleware/archive.js
@@ -1,31 +1,8 @@
 'use strict';
 
-const crypto = require('crypto');
 const Raven = require('raven');
 const axios = require('axios').default;
-
-const pathToFilename = path => {
-	// Same name regardless of host.
-	// https://origami-build.ft.com/[...]
-	// https://www.ft.com/__origami/service/build/[...]
-	const pathWithoutOrigin = path.replace(/^\/?__origami\/service\/build\/?/, '').replace(/^\//, '');
-	// Decode for human readability.
-	const decodedPathWithoutOrigin = decodeURIComponent(pathWithoutOrigin);
-	return decodedPathWithoutOrigin
-		// Responses for one endpoint under one directory.
-		// File for query param combination.
-		.replace(/\?/, '/?')
-		// Sanitise url. Keep select special characters for
-		// human readability.
-		.replace(/[^a-zA-Z0-9-_^,&?/]/g, '_')
-		// Remove any slash from a query param (we decoded the url, and don't want
-		// someone to manipulate the structure of our s3 bucket with a query param).
-		.replace(/(?:[?])(?:.+)?[/]/g, '_')
-		// Append hash to avoid conflicts having crudely sanitised the url.
-		// Use hash of decoded path always.
-		+ '-' +
-	crypto.createHash('md5').update(decodedPathWithoutOrigin).digest('hex');
-};
+const pathToFilename = require('../path-to-archive-filename');
 
 module.exports = function (app) {
 	const {metrics} = app.ft;
