+++ test/integration/v2-archive.test.js
@@ -68,7 +68,8 @@ describe('Archived routes', function() {
 					before(async function () {
 						const now = (new Date()).toISOString();
 						response = await request(this.app)
-							.get(`/v2/bundles/css?modules=o-card@%5E3.0.0&test-archive-miss=${now}`)
+							// export used to cache bust, unrecognised params are removed by the archive
+							.get(`/v2/bundles/css?modules=o-card@%5E3.0.0&export=${now.replace(/[^a-zA-Z0-9]/g, '')}`)
 							.redirects(5)
 							.set('Connection', 'close');
 					});
@@ -131,7 +132,8 @@ describe('Archived routes', function() {
 					before(async function () {
 						const now = (new Date()).toISOString();
 						response = await request(this.app)
-							.get(`/v2/bundles/js?modules=o-tracking&test-archive-miss=${now}`)
+							// export used to cache bust, unrecognised params are removed by the archive
+							.get(`/v2/bundles/js?modules=o-tracking&export=${now.replace(/[^a-zA-Z0-9]/g, '')}`)
 							.redirects(5)
 							.set('Connection', 'close');
 					});
@@ -173,7 +175,8 @@ describe('Archived routes', function() {
 					before(async function () {
 						const now = (new Date()).toISOString();
 						response = await request(this.app)
-							.get(`/v2/files/o-comments@3.5.0/src/images/comment_featured_close_quote.png?test-archive-miss=${now}`)
+							// export used to cache bust, unrecognised params are removed by the archive
+							.get(`/v2/files/o-comments@3.5.0/src/images/comment_featured_close_quote.png?export=${now.replace(/[^a-zA-Z0-9]/g, '')}`)
 							.redirects(5)
 							.set('Connection', 'close');
 					});
@@ -249,7 +252,8 @@ describe('Archived routes', function() {
 					before(async function () {
 						const now = (new Date()).toISOString();
 						response = await request(this.app)
-							.get(`/v2/bundles/css?modules=o-card@%5E3.0.0&test-archive-miss=${now}`)
+							// export used to cache bust, unrecognised params are removed by the archive
+							.get(`/v2/bundles/css?modules=o-card@%5E3.0.0&export=${now.replace(/[^a-zA-Z0-9]/g, '')}`)
 							.redirects(5)
 							.set('Connection', 'close');
 					});
@@ -327,7 +331,8 @@ describe('Archived routes', function() {
 					before(async function () {
 						const now = (new Date()).toISOString();
 						response = await request(this.app)
-							.get(`/v2/bundles/js?modules=o-tracking&test-archive-miss=${now}`)
+							// export used to cache bust, unrecognised params are removed by the archive
+							.get(`/v2/bundles/js?modules=o-tracking&export=${now.replace(/[^a-zA-Z0-9]/g, '')}`)
 							.redirects(5)
 							.set('Connection', 'close');
 					});
@@ -380,7 +385,8 @@ describe('Archived routes', function() {
 					before(async function () {
 						const now = (new Date()).toISOString();
 						response = await request(this.app)
-							.get(`/v2/files/o-comments@3.5.0/src/images/comment_featured_close_quote.png?test-archive-miss=${now}`)
+							// export used to cache bust, unrecognised params are removed by the archive
+							.get(`/v2/files/o-comments@3.5.0/src/images/comment_featured_close_quote.png?export=${now.replace(/[^a-zA-Z0-9]/g, '')}`)
 							.redirects(5)
 							.set('Connection', 'close');
 					});
